Implemented dynamic/infinite scrolling for tileset viewer, metatiles viewer, and level maps.
Scrolling is now done in a similar style to this new "hex gamma" hex viewer structure. Still a few bugs and quirks but it seems to improve performance, at least now full level maps can be viewed. More efficient storage of tile and metatile canvas elements, and reduced number of canvas elements used in a metatile (now just 1 instead of 16). Could be optimized further still. Bitplane viewer was fixed slightly, although it may be phased out in the future anyway.